Big refacto of `DistillationTrainer`: - Create one class per distil method - Deprecate TACDistillation
